+++ packages/common/src/controllers/AccountController.ts
@@ -6,7 +6,9 @@ import type { IntegrationType } from '../../types/config';
 import CheckoutService from '../services/integrations/CheckoutService';
 import AccountService, { type AccountServiceFeatures } from '../services/integrations/AccountService';
 import SubscriptionService from '../services/integrations/SubscriptionService';
+import JWPEntitlementService from '../services/JWPEntitlementService';
 import type { Offer } from '../../types/checkout';
+import type { Plan } from '../../types/plans';
 import type {
   Capture,
   Customer,
@@ -16,7 +18,7 @@ import type {
   GetCaptureStatusResponse,
   SubscribeToNotificationsPayload,
 } from '../../types/account';
-import { assertFeature, assertModuleMethod, getNamedModule } from '../modules/container';
+import { assertFeature, assertModuleMethod, getModule, getNamedModule } from '../modules/container';
 import { INTEGRATION_TYPE } from '../modules/types';
 import type { ServiceResponse } from '../../types/service';
 import { useAccountStore } from '../stores/AccountStore';
@@ -26,12 +28,15 @@ import { logError } from '../logger';
 
 import WatchHistoryController from './WatchHistoryController';
 import FavoritesController from './FavoritesController';
+import AccessController from './AccessController';
 
 @injectable()
 export default class AccountController {
   private readonly checkoutService: CheckoutService;
   private readonly accountService: AccountService;
   private readonly subscriptionService: SubscriptionService;
+  private readonly entitlementService: JWPEntitlementService;
+  private readonly accessController: AccessController;
   private readonly favoritesController: FavoritesController;
   private readonly watchHistoryController: WatchHistoryController;
   private readonly features: AccountServiceFeatures;
@@ -41,14 +46,17 @@ export default class AccountController {
 
   constructor(
     @inject(INTEGRATION_TYPE) integrationType: IntegrationType,
+    accessController: AccessController,
     favoritesController: FavoritesController,
     watchHistoryController: WatchHistoryController,
   ) {
     this.checkoutService = getNamedModule(CheckoutService, integrationType);
     this.accountService = getNamedModule(AccountService, integrationType);
     this.subscriptionService = getNamedModule(SubscriptionService, integrationType);
+    this.entitlementService = getModule(JWPEntitlementService);
 
     // @TODO: Controllers shouldn't be depending on other controllers, but we've agreed to keep this as is for now
+    this.accessController = accessController;
     this.favoritesController = favoritesController;
     this.watchHistoryController = watchHistoryController;
 
@@ -85,6 +93,7 @@ export default class AccountController {
     useConfigStore.setState({ accessModel: this.accountService.accessModel });
 
     await this.loadUserData();
+    await this.getEntitledPlans();
 
     useAccountStore.setState({ loading: false });
   };
@@ -163,6 +172,7 @@ export default class AccountController {
       const response = await this.accountService.login({ email, password, referrer });
 
       if (response) {
+        await this.accessController?.generateAccessTokens();
         await this.afterLogin(response.user, response.customerConsents);
         return;
       }
@@ -180,6 +190,7 @@ export default class AccountController {
 
   logout = async () => {
     await this.accountService?.logout();
+    await this.accessController?.removeAccessTokens();
     await this.clearLoginState();
 
     // let the application know to refresh all entitlements
@@ -378,6 +389,29 @@ export default class AccountController {
 
     const { responseData } = await this.checkoutService.getEntitlements({ offerId });
     return !!responseData?.accessGranted;
+  };
+
+  // This currently supports only one plan, as the current usage for the media metadata requires only one plan_id provided.
+  // TODO: Support for multiple plans should be added. Revisit this logic once the dependency on plan_id is changed.
+  getEntitledPlans = async (): Promise<Plan | null> => {
+    const { config, settings } = useConfigStore.getState();
+    const siteId = config.siteId;
+    const isAccessBridgeEnabled = !!settings?.apiAccessBridgeUrl;
+
+    // This should be only used when access bridge is defined, regardless of the integration type.
+    if (!isAccessBridgeEnabled) {
+      return null;
+    }
+
+    const response = await this.entitlementService.getEntitledPlans({ siteId });
+    if (response?.plans?.length) {
+      // Find the SVOD plan or fallback to the first available plan
+      const entitledPlan = response.plans.find((plan) => plan.metadata.access_model === 'svod') || response.plans[0];
+      useAccountStore.setState({ entitledPlan });
+      return entitledPlan;
+    }
+
+    return null;
   };
 
   reloadSubscriptions = async (