+++ packages/common/src/modules/register.ts
@@ -3,7 +3,7 @@
 import 'reflect-metadata'; // include once in the app for inversify (see: https://github.com/inversify/InversifyJS/blob/master/README.md#-installation)
 import { INTEGRATION, EPG_TYPE } from '../constants';
 import { container } from './container';
-import { DETERMINE_INTEGRATION_TYPE, INTEGRATION_TYPE } from './types';
+import { API_ACCESS_BRIDGE_URL, DETERMINE_INTEGRATION_TYPE, INTEGRATION_TYPE } from './types';
 
 import ApiService from '../services/ApiService';
 import WatchHistoryService from '../services/WatchHistoryService';
@@ -16,6 +16,7 @@ import SettingsService from '../services/SettingsService';
 import WatchHistoryController from '../controllers/WatchHistoryController';
 import CheckoutController from '../controllers/CheckoutController';
 import AccountController from '../controllers/AccountController';
+import AccessController from '../controllers/AccessController';
 import FavoritesController from '../controllers/FavoritesController';
 import AppController from '../controllers/AppController';
 import EpgController from '../controllers/EpgController';
@@ -24,6 +25,10 @@ import EpgController from '../controllers/EpgController';
 import EpgService from '../services/EpgService';
 import ViewNexaEpgService from '../services/epg/ViewNexaEpgService';
 import JWEpgService from '../services/epg/JWEpgService';
+
+// Access integration
+import AccessService from '../services/AccessService';
+import { getApiAccessBridgeUrl } from './functions/getApiAccessBridgeUrl';
 
 // Integration interfaces
 import AccountService from '../services/integrations/AccountService';
@@ -51,6 +56,7 @@ container.bind(FavoriteService).toSelf();
 container.bind(GenericEntitlementService).toSelf();
 container.bind(ApiService).toSelf();
 container.bind(SettingsService).toSelf();
+container.bind(AccessService).toSelf();
 
 // Common controllers
 container.bind(AppController).toSelf();
@@ -61,6 +67,7 @@ container.bind(EpgController).toSelf();
 // Integration controllers
 container.bind(AccountController).toSelf();
 container.bind(CheckoutController).toSelf();
+container.bind(AccessController).toSelf();
 
 // EPG services
 container.bind(EpgService).to(JWEpgService).whenTargetNamed(EPG_TYPE.jwp);
@@ -68,6 +75,7 @@ container.bind(EpgService).to(ViewNexaEpgService).whenTargetNamed(EPG_TYPE.viewN
 
 // Functions
 container.bind(INTEGRATION_TYPE).toDynamicValue(getIntegrationType);
+container.bind(API_ACCESS_BRIDGE_URL).toDynamicValue(getApiAccessBridgeUrl);
 
 // Cleeng integration
 container.bind(DETERMINE_INTEGRATION_TYPE).toConstantValue(isCleengIntegrationType);